Modify shuttle schema, add cancel and delete shuttle.
Allow shuttles to now have guests, and to set the maximum number of guests allowed per user.  Allow shuttles to have a maxCapacity (because, that's pretty important). Allow admins to now cancel and delete shuttles. Cancelling a shuttle keeps it in the database, but sets its active flag to false (disabling signups for it). Delete shuttle will permanently remove a shuttle from the database (without a trace). I don't know why an admin would want to delete a shuttle (unless making it was a total mistake, and no one has even signed up for it).